add bannedWords whitelist

+++ index.js
@@ -631,6 +631,10 @@ client.on("guildMemberRemove", member => {
 	delete client.userLevels._content[member.user.id];
 });
 
+// whitelisted words
+client.whitelistWords = new database("./databases/whitelistWords.json", "array");
+client.whitelistWords.initLoad();
+
 // banned words
 client.bannedWords = new database("./databases/bannedWords.json", "array");
 client.bannedWords.initLoad();
@@ -810,7 +814,7 @@ client.on("messageCreate", async (message) => {
 		message.reply(":b:ingus y u use userbenchmark");
 	}
 	// handle banned words
-	if (client.bannedWords._content.some(word => message.content.toLowerCase().includes(word)) && message.channel.id !== (client.config.mainServer.channels.modchat) && message.guild.id === client.config.mainServer.id) {
+	if (client.bannedWords._content.some(word => message.content.toLowerCase().includes(word)) && (!client.whitelistWords._content.some(word => message.content.toLowerCase().includes(word))) && message.channel.id !== (client.config.mainServer.channels.modchat) && message.guild.id === client.config.mainServer.id) {
 		message.delete()
 		message.channel.send("That word is banned here.").then(x => setTimeout(() => x.delete(), 5000))}
 });
